pyproject.toml
[project]
name = "wisey"
version = "0.1.0"
description = "Thinkwise Knowledge Agent — RAG-based Q&A from docs, community, and notes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.6",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "psycopg2-binary>=2.9.11",
    "tiktoken>=0.12.0",
]
